+++ crypto/schnorr-adaptor-signature/keys_test.go
@@ -20,12 +20,14 @@ func FuzzKeyGen(f *testing.F) {
 		require.NoError(t, err)
 
 		// ensure that decKey.GetEncKey() is same as encKey
-		actualEncKey := decKey.GetEncKey()
+		actualEncKey, err := decKey.GetEncKey()
+		require.NoError(t, err)
 		require.Equal(t, encKey, actualEncKey)
 
 		// ensure that the corresponding btcPK and btcSK
 		// constitute a key pair
-		btcPK := encKey.ToBTCPK()
+		btcPK, err := encKey.ToBTCPK()
+		require.NoError(t, err)
 		btcSK := decKey.ToBTCSK()
 		actualBTCPK := btcSK.PubKey()
 		require.Equal(t, btcPK, actualBTCPK)
@@ -35,7 +37,7 @@ func FuzzKeyGen(f *testing.F) {
 		actualEncKey, err = asig.NewEncryptionKeyFromBTCPK(btcPK)
 		require.NoError(t, err)
 		require.Equal(t, encKey, actualEncKey)
-		actualDecKey, err := asig.NewDecyptionKeyFromBTCSK(btcSK)
+		actualDecKey, err := asig.NewDecryptionKeyKeyFromBTCSK(btcSK)
 		require.NoError(t, err)
 		require.Equal(t, decKey, actualDecKey)
 	})
@@ -61,7 +63,7 @@ func FuzzKeySerialization(f *testing.F) {
 
 		// roundtrip of serialising/deserialising decKey
 		decKeyBytes := decKey.ToBytes()
-		actualDecKey, err := asig.NewDecyptionKeyFromBytes(decKeyBytes)
+		actualDecKey, err := asig.NewDecryptionKeyKeyFromBytes(decKeyBytes)
 		require.NoError(t, err)
 		require.Equal(t, decKey, actualDecKey)
 	})