+++ CHANGELOG.md
@@ -85,6 +85,7 @@ invocation
 
 ### Bug fixes
 
+- [#657](https://github.com/babylonlabs-io/babylon/pull/657) crypto: fix adaptor sig timing side channels
 - [#591](https://github.com/babylonlabs-io/babylon/pull/591) bump ibc to v8.6.1 that fixes security issue
 - [#579](https://github.com/babylonlabs-io/babylon/pull/579) Slashed FP gets activated in voting power distribution
 cache if an old BTC delegation receives inclusion proof